Move utilities.utils.render_jinja2() to utilities.jinja2
