+++ netbox/extras/models/models.py
@@ -24,7 +24,8 @@ from netbox.models.features import (
 )
 from utilities.html import clean_html
 from utilities.querysets import RestrictedQuerySet
-from utilities.utils import dict_to_querydict, render_jinja2
+from utilities.jinja2 import render_jinja2
+from utilities.utils import dict_to_querydict
 
 __all__ = (
     'Bookmark',
